requirements.txt
streamlit
google-generativeai
fpdf
matplotlib
python-dotenv
